refactor: moved HttpProxy to standalone file
